refactor: 미팅, 미팅 상세 페이지 폴더 구조 재정의
