+++ next.config.ts
@@ -8,6 +8,8 @@ const nextConfig: NextConfig = {
       "example.com",
       "lh3.googleusercontent.com",
       "sprint-fe-project.s3.ap-northeast-2.amazonaws.com",
+      "source.unsplash.com",
+      "picsum.photos",
     ],
   },
 };
