+++ src/api/meetings.ts
@@ -8,9 +8,7 @@ import {
   MyMeetingsResponse,
   FavoritesResponse,
 } from "@/types";
-import type { MeetingParticipantsResponse } from "@/app/meetings/[meetingId]/types";
-
-
+import type { MeetingParticipantsResponse } from "@/types/meetingTypes";
 
 export async function getMeetingList(
   params: GetMeetingListParams,
@@ -52,7 +50,6 @@ export async function updateMeeting(
   const { data } = await axiosInstance.patch(`/meetings/${meetingId}`, params);
   return data;
 }
-
 
 export async function updateFavorites(meetingId: number): Promise<void> {
   await axiosInstance.post(`/meetings/${meetingId}/favorites`);
