--- src/app/meetings/[meetingId]/model/meeting-detail.errors.ts
@@ -1,29 +1,0 @@
-export const getJoinErrorMessage = (code?: string) => {
-  switch (code) {
-    case "CANCELED":
-      return "취소된 모임은 참여할 수 없어요.";
-    case "REGISTRATION_CLOSED":
-      return "모집이 마감된 모임이에요.";
-    case "CAPACITY_FULL":
-      return "정원이 가득 찬 모임이에요.";
-    case "ALREADY_JOINED":
-      return "이미 참여 중인 모임이에요.";
-    case "NOT_FOUND":
-      return "존재하지 않는 모임이에요.";
-    case "REFRESH_FAILED":
-      return "로그인 후 다시 참여해 주세요.";
-    default:
-      return "참여 처리 중 문제가 발생했어요.";
-  }
-};
-
-export const getCancelJoinErrorMessage = (code?: string) => {
-  switch (code) {
-    case "NOT_FOUND":
-      return "존재하지 않는 모임이에요.";
-    case "REFRESH_FAILED":
-      return "로그인 후 다시 참여 취소해 주세요.";
-    default:
-      return "참여 취소 처리 중 문제가 발생했어요.";
-  }
-};
