+++ src/app/meetings/modal/MeetingBasicInfoSection.tsx
@@ -1,7 +1,5 @@
 "use client";
 
-import { DEFAULT_MEETING_TYPE_OPTIONS } from "@/app/meetings/modal/model/meeting-form.constants";
-import { MeetingBasicInfoSectionProps } from "@/app/meetings/modal/model/meeting-form.props";
 import { ImageUploadInput } from "@/components/ui/ImageUploadInput";
 import {
   Select,
@@ -12,6 +10,61 @@ import {
 } from "@/components/shadcnOrigin/select";
 import { InputCommon } from "@/components/ui/InputCommon";
 import { TextareaCommon } from "@/components/ui/TextareaCommon";
+
+import type { StaticImageData } from "next/image";
+import teamBulb from "@/assets/img/bulb/elec-bulb.jpg";
+import studyImage from "@/assets/img/category/study.jpg";
+import towerWork from "@/assets/img/category/business.jpg";
+import JobIShoes from "@/assets/img/category/fitness-health.jpg";
+import etcImage from "@/assets/img/category/etc.jpg";
+import { MeetingBasicInfoSectionProps } from "@/types/meeting/meeting-form.props";
+
+export interface MeetingTypeOption {
+  value: string;
+  label: string;
+}
+
+export interface MeetingCategoryItem {
+  value: string;
+  label: string;
+  imageSrc: StaticImageData;
+  className?: string;
+}
+
+export const MEETING_CATEGORY_LIST: MeetingCategoryItem[] = [
+  {
+    value: "친목/여가",
+    label: "친목/여가",
+    imageSrc: teamBulb,
+  },
+  {
+    value: "스터디",
+    label: "스터디",
+    imageSrc: studyImage,
+  },
+  {
+    value: "워케이션",
+    label: "워케이션",
+    imageSrc: towerWork,
+  },
+  {
+    value: "취미/운동",
+    label: "취미/운동",
+    imageSrc: JobIShoes,
+  },
+  {
+    value: "기타",
+    label: "기타",
+    imageSrc: etcImage,
+    className: "col-span-2",
+  },
+];
+
+export const DEFAULT_MEETING_TYPE_OPTIONS: MeetingTypeOption[] =
+  MEETING_CATEGORY_LIST.map(({ value, label }) => ({
+    value,
+    label,
+  }));
 
 export function MeetingBasicInfoSection({
   values,
