+++ src/app/meetings/modal/MeetingScheduleStep.tsx
@@ -1,9 +1,9 @@
 "use client";
 
-import { MeetingScheduleStepProps } from "@/app/meetings/modal/model/meeting-form.props";
 import { ScheduleDatePicker } from "@/app/meetings/modal/ScheduleDatePicker";
 import { ScheduleTimePicker } from "@/app/meetings/modal/ScheduleTimePicker";
 import { InputCommon } from "@/components/ui/InputCommon";
+import { MeetingScheduleStepProps } from "@/types/meeting/meeting-form.props";
 
 const getTodayDateString = () => {
   const now = new Date();
