+++ src/app/my-meetings/page.tsx
@@ -1,7 +1,5 @@
 import Image from "next/image";
-import {
-  InfiniteData,
-} from "@tanstack/react-query";
+import { InfiniteData } from "@tanstack/react-query";
 
 import savedLg from "@/assets/img/head/saved-lg.jpg";
 import savedSm from "@/assets/img/head/saved-sm.jpg";
@@ -11,11 +9,9 @@ import MyMeetingsClient from "./components/MyMeetingsClient";
 import PrefetchBoundary from "@/components/boundary/PrefetchBoundary";
 import { Suspense } from "react";
 import MeetingCardSkeleton from "@/components/skeleton/MeetingCardSkeleton";
-import { getMeetingJoinedNextPageParam } from "@/hooks/useMeetingQuery";
-
+import { getMeetingJoinedNextPageParam } from "@/hooks/meetings/useMeetingQuery";
 
 export default async function Page() {
-
   return (
     <div className="w-full bg-gray-50 pt-6 pb-20 sm:pt-10 lg:pt-[48px]">
       <div className="mx-auto w-full max-w-[1280px] px-4 sm:px-6 lg:px-8">
