+++ src/types/meeting/meeting-image.types.ts
